Day 8

+++ com/learn/java/ch3/VariablesIntro.java
@@ -2,14 +2,37 @@ package com.learn.java.ch3;
 
 public class VariablesIntro {
 	public static void main(String[] args) {
-		int a;
+		
+		
+		byte a;
+		//Declaring a variable named a
+		
 		a=10;
+		//Initializing the variable a with value 10
+		
+		
 		System.out.println(a);
-		int b=20;
+		//Printing variable a
+
+		
+		byte b=20;
+		//Declaring and initializing a variable named b with value 20
+		
+		
 		System.out.println(b);
-		int c;
+		//Printing variable b
+		
+		
+		byte c;
+		//Declaring a variable named c
+		
+		
 		c=a;
+		//Initializing the variable c with variable a
+		
+		
 		System.out.println(c);
+		//Printing variable c
 	}
 
 }
